+++ modules/ActCommon/mod.ts
@@ -4,4 +4,5 @@ export * from './useAsync.ts';
 export * from './useDisposable.ts';
 export * from './isArrayEqual.ts';
 export * from './useSelection.ts';
+export * from './useBlobURL.ts';
 export { curve } from './deps.ts';