+++ modules/ActCommon/useAsync.ts
@@ -5,19 +5,16 @@ const { useState } = act;
 
 export const useAsync = <T>(createPromise: () => Promise<T>, deps: act.Deps = []): T | null => {
   const [resolution, setResolution] = useState<T | null>(null);
-  const [rejection, setRejection] = useState<unknown | null>(null);
+  const [, setRejection] = useState<unknown | null>(null);
 
   useEffect(() => {
     createPromise()
       .then(setResolution)
-      .catch(setRejection)
+      .catch(e => (console.error(e), setRejection(e)))
   }, deps);
 
   if (resolution)
     return resolution;
-
-  if (rejection)
-    throw rejection;
 
   return null;
 }