+++ modules/Clerk/ui/deps.ts
@@ -1,2 +1,3 @@
 export * as threeCommon from '../../ThreeCommon/mod.ts';
 export { actThree, act } from '../../AtlasRenderer/mod.ts'
+export * as artifact from '../../Artifact/mod.ts';