+++ modules/ComponentDoc/deps.ts
@@ -1,3 +1,4 @@
 export { act } from '../AtlasRenderer/mod.ts';
 export * as actNavigation from "https://esm.sh/@lukekaalim/act-navigation@1.2.1";
 export * as janitor from '../Janitor/mod.ts';
+export * as actMarkdown from "https://esm.sh/@lukekaalim/act-markdown@1.8.1";