+++ modules/ComponentDoc/mod.ts
@@ -7,3 +7,8 @@ export * from './markdown.ts';
 export * from './Blocks.ts';
 export * from './DocContext.ts';
 export * from './DocSite2.ts';
+export * from './global.ts';
+
+import { actMarkdown, act } from './deps.ts';
+
+export type DirectiveComponent = act.Component<actMarkdown.MarkdownDirectiveComponentProps>;