+++ modules/ThreeCommon/deps.ts
@@ -1,5 +1,6 @@
 export * as schedule from '../FrameScheduler/mod.ts';
 export { act, actThree, three } from '../AtlasRenderer/mod.ts';
+export * as actCommon from '../ActCommon/mod.ts';
 
 // @deno-types="https://esm.sh/v131/@types/three@0.155.1/examples/jsm/renderers/CSS2DRenderer~.d.ts"
 export * as css2d from 'https://esm.sh/three@0.155.0/examples/jsm/renderers/CSS2DRenderer';
@@ -8,3 +9,5 @@ export * as gltf from 'https://esm.sh/three@0.155.0/examples/jsm/loaders/GLTFLoa
 
 
 export * as rxjs from "https://esm.sh/rxjs@7.8.1";
+
+export * as kayo from '../Kayo/mod.ts';