+++ modules/Wizard/asset/props.ts
@@ -1,0 +1,3 @@
+export type WizardAssetProps = {
+  content: Blob,
+}