+++ modules/Wizard/deps.ts
@@ -1,2 +1,8 @@
-export { act } from '../AtlasRenderer/mod.ts';
-export * as journal from '../Journal/mod.ts';
+export { act, three, actThree } from '../AtlasRenderer/mod.ts';
+
+export * as actCommon from '../ActCommon/mod.ts';
+export * as threeCommon from '../ThreeCommon/mod.ts';
+
+export * as keyboard from '../Keyboard/mod.ts';
+export * as deskplane from '../Deskplane/mod.ts';
+export * as formula from '../Formula/mod.ts';