+++ modules/Wizard/docs.ts
@@ -1,0 +1,8 @@
+import { doc } from "../ComponentDoc/mod.ts";
+
+import "./asset/docs.ts";
+
+doc({
+  id: 'Wizard',
+  readmeURL: new URL('./readme.md', import.meta.url),
+});