+++ modules/Wizard/managers/wizardManager.ts
@@ -1,9 +1,0 @@
-import { journal } from "../deps.ts";
-
-export type WizardManager = {
-  game: journal.GameController,
-};
-
-export const createWizardManager = () => {
-
-};